requirements.txt
streamlit
numpy
opencv-python
networkx
torch
torchvision
Pillow
scikit-image
scikit-learn
tqdm 